Issue #48: per Alex directive, moved (de)serialise() utility functions to SchemaUtil. Decommissioned UtilIO.
